Modified data type

+++ Efficient Fibonacci series Algo.cpp
@@ -2,11 +2,11 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-vector<long> fib(int n)
+vector<long long> fib(int n)
 {
 	int f=0;
 	int s=1;
-vector<long> v;
+vector<long long> v;
 v.push_back(f);
 v.push_back(s);
 	for(int i=2;i<=n;i++)
@@ -20,7 +20,7 @@ int main()
 	cout<<"enter the number for calc Fibonacci series :"<<endl;
 	int n;
 	cin>>n;
-	vector<long> v=fib(n);
+	vector<long long> v=fib(n);
 	cout<<"Fib series is :"<<endl;
     for(int i=0;i<n;i++)
     cout<<v[i]<<" ";
